third_party/WebKit/LayoutTests/platform/android/fast/notifications/notifications-cancel-request-permission-expected.txt

CONSOLE ERROR: line 16: Uncaught TypeError: Object [object Object] has no method 'ignoreLegacyWebNotificationPermissionRequests'
Requesting and cancelling notification permission...

Pass if notification is cancelled
